move AssistantBuilderDataSourceModal to actions/configuration
